revert content table
